// features/markdown/types/markdown.ts
export type MarkdownFormatAction =
  | "bold"
  | "italic"
  | "strikethrough"
  | "quote"
  | "titleCase"
  | "upperCase"
  | "lowerCase"
  | "h1"
  | "h2"
  | "h3"
  | "h4"
  | "h5"
  | "h6"
  | "alignLeft"
  | "alignCenter"
  | "alignRight"
  | "orderedList"
  | "bulletList"
  | "taskList"
  | "horizontalRule"
  | "inlineCode"
  | "codeBlock"
  | "link"
  | "reference"
  | "image"
  | "table"
  | "math"
  | "emoji"
  | "symbol"
  | "alertNote"
  | "alertTip"
  | "alertImportant"
  | "alertWarning"
  | "alertCaution";

export type MarkdownViewMode = "editor" | "split" | "preview";